Create PLACE_SWITRS_DB_HERE
Replace this .txt file with a freshly decompressed copy of the switrs.sqite source file after cloning this repository on your local computer. 

The SWITRS database source file is far to large to manage in a typical free-tier GitHub repository. Please download the switrs.sqlite source file from Kaggle.com and integrate it with the cloned contents of this repository.

-Micah614